+++ src/present.cpp
@@ -55,7 +55,7 @@ namespace libsemigroups {
 
     word_type prod(std::initializer_list<size_t> ilist,
                    size_t                        first,
-                   size_t                        last,
+                   int                           last,
                    size_t                        step) {
       return prod(word_type(ilist), first, last, step);
     }
